refactor: move state propagation code from customer portal to store mixin
